+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+* Upgrade to Swift 5.6 - [@Blackjacx](https://github.com/blackjacx).
 
 ## [0.0.3] - 2022-03-14Z
 * [#3](https://github.com/blackjacx/engine/pull/3): Async Await Support - [@Blackjacx](https://github.com/blackjacx).